% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Q_plot.R
\name{create_Q_plot}
\alias{create_Q_plot}
\title{Plot Q function based on mone carlo estimation}
\usage{
create_Q_plot(X, Y, k_plot_grid = 100, MC = 100, print = TRUE)
}
\arguments{
\item{X}{numerical vector with first random variable e.g. c(1.1, 2.2, 1.73).}

\item{Y}{numerical vector with second random variable e.g. c(3.1, 1.2, 1.93).}

\item{k_plot_grid}{number of grid points for plot e.g. 1000, defaults to 100.}

\item{MC}{number of replications e.g. 1000, defaults to 100.}

\item{print}{boolean value, if true a plot will be displayed, defaults to TRUE.}
}
\value{
A character string, capitalized to title case.
}
\description{
Plot Q function based on mone carlo estimation
}
\examples{
'calculate_copula_mc_grid(c(1, 3, 2), c(3, 2, 1)) '
}
