\name{VoA-package}
\alias{VoA-package}
\alias{VoA}
\docType{package}
\title{
\packageTitle{VoA}
}
\description{
\packageDescription{VoA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VoA}
\packageIndices{VoA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VoA}

Maintainer: \packageMaintainer{VoA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
